requirements.txt
APScheduler==3.10.1
async-timeout==4.0.2
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
filelock==3.12.0
Flask==2.2.3
Flask-APScheduler==1.12.4
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.15.0
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mysql-connector-python==8.0.33
mysqlclient==2.1.1 
opencv-python-headless
packaging==23.1
platformdirs==3.2.0
protobuf==3.20.3
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
redis==4.5.4
six==1.16.0
SQLAlchemy==2.0.10
tensorflow==2.12.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
virtualenv==20.22.0
Werkzeug==2.2.3
